Correction to PR's comment
